feat: Allow posting multiple images
This adds the Gallery component to the editor for Photo posts. Since the
editor only deals with a single photo at once, the alt data is now
stored in cache instead alongside the URL.

The order of the images can be changed using the arrows, and a picture
removed using the X. This doesn't remove it from the cache, so
refreshing the gallery will bring it back (presuming it's still in the
server response).

This completely trashes the close button on the New Photo post, however,
as changing between photos updates the URL, and the close button is just
moving back, I think. Ah well.